requirements.txt
# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
fake-useragent==1.4.0

# Data processing
pandas==2.1.3

# Database - PostgreSQL
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0

# Development
jupyter==1.0.0
